fix: fixed typos
